+++ src/font.h
@@ -4,7 +4,7 @@
 #include "math.h"
 #include "atlas.h"
 
-void RenderTextColor(const Atlas *atlas, const char *text, i32 x, i32 y, Color color);
-void RenderText(const Atlas *atlas, const char *text, i32 x, i32 y);
+void RenderTextColor(const Atlas *atlas, const char *text, float x, float y, Color color);
+void RenderText(const Atlas *atlas, const char *text, float x, float y);
 
 #endif
